chore: make tests for `acvm::pwg::solve` integration tests
